+++ src/hardis-colors.ts
@@ -57,7 +57,7 @@ export class HardisColors {
       for (const sfdxConfigPath of this.sfdxConfigPaths) {
         const sfdxConfigFullPath = path.join(
           vscode.workspace.workspaceFolders[0].uri.fsPath,
-          sfdxConfigPath,
+          sfdxConfigPath
         );
         if (fs.existsSync(sfdxConfigFullPath)) {
           await this.manageColor(sfdxConfigFullPath);
@@ -74,8 +74,8 @@ export class HardisColors {
         const watcher = vscode.workspace.createFileSystemWatcher(
           new vscode.RelativePattern(
             vscode.workspace.workspaceFolders[0],
-            `**/${sfdxConfigPath}`,
-          ),
+            `**/${sfdxConfigPath}`
+          )
         );
         watcher.onDidCreate((uri) => this.manageColor(uri.fsPath));
         watcher.onDidChange((uri) => this.manageColor(uri.fsPath));
@@ -102,10 +102,10 @@ export class HardisColors {
         } else {
           vscode.window.showWarningMessage(
             "🦙 You need to select a default org to define a color for it :)",
-            "Close",
+            "Close"
           );
         }
-      },
+      }
     );
     this.disposables.push(disposable);
   }
@@ -139,7 +139,7 @@ export class HardisColors {
     if (fileDefaultOrg !== this.currentDefaultOrg) {
       this.currentDefaultOrg = fileDefaultOrg;
       this.currentDefaultOrgDomain = await getUsernameInstanceUrl(
-        this.currentDefaultOrg || "",
+        this.currentDefaultOrg || ""
       );
       const orgColor = await this.getCurrentDefaultOrgColor();
       await this.applyColor(orgColor);
@@ -147,7 +147,7 @@ export class HardisColors {
       if (this.initializing === false) {
         vscode.commands.executeCommand(
           "vscode-sfdx-hardis.refreshStatusView",
-          true,
+          true
         );
       }
     }
@@ -169,28 +169,28 @@ export class HardisColors {
       {
         fail: false,
         output: true,
-      },
+      }
     );
     if (orgRes?.result?.records?.length === 1) {
       const org = orgRes.result.records[0];
       if (org.IsSandbox === true) {
         // We are in a dev sandbox or scratch org !
         const isMajorOrg = await this.isMajorOrg(
-          this.currentDefaultOrgDomain || "",
+          this.currentDefaultOrgDomain || ""
         );
         if (isMajorOrg) {
           vscode.window.showWarningMessage(
             `🦙 Your default org is a MAJOR org linked to git branch ${this.majorOrgBranch}, be careful because the CI/CD Server is supposed to deploy there, not you :)`,
-            "Close",
+            "Close"
           );
           return forcedColor || "#a66004"; // orange !
         }
-        return forcedColor || "#04590c"; // green !
+        return forcedColor || null; // green !
       } else if (PRODUCTION_EDITIONS.includes(org.OrganizationType)) {
         // We are in production !!
         vscode.window.showWarningMessage(
           "🦙 Your default org is a PRODUCTION org, be careful what you do :)",
-          "Close",
+          "Close"
         );
         return forcedColor || "#8c1004"; // red !
       }
@@ -206,17 +206,31 @@ export class HardisColors {
     if (vscode.workspace.workspaceFolders) {
       const config = vscode.workspace.getConfiguration();
       const colorCustomization: any = config.get(
-        "workbench.colorCustomizations",
+        "workbench.colorCustomizations"
       );
-      colorCustomization["statusBar.background"] = color || undefined;
-      colorCustomization["activityBar.background"] = color || undefined;
-      // Do not update config file with blank color if there wasn't a previous config
-      if (color || Object.keys(colorCustomization).length > 0) {
+      if (color !== null) {
+        colorCustomization["statusBar.background"] = color;
+        colorCustomization["activityBar.background"] = color;
+        // Do not update config file with blank color if there wasn't a previous config
         await config.update(
           "workbench.colorCustomizations",
           colorCustomization,
-          vscode.ConfigurationTarget.Workspace,
+          vscode.ConfigurationTarget.Workspace
         );
+      } else if (
+        colorCustomization["statusBar.background"] ||
+        colorCustomization["activityBar.background"]
+      ) {
+        // Remove colors
+        colorCustomization["statusBar.background"] = null;
+        colorCustomization["activityBar.background"] = null;
+        if (Object.keys(colorCustomization).length > 0) {
+          await config.update(
+            "workbench.colorCustomizations",
+            {},
+            vscode.ConfigurationTarget.Workspace
+          );
+        }
       }
     }
   }
@@ -225,9 +239,9 @@ export class HardisColors {
     this.majorOrgBranch = undefined;
     const majorOrgInstanceUrls = await this.listMajorOrgsInstanceUrls();
     const matchOrgs = majorOrgInstanceUrls.filter(
-      (org) => org.instanceUrl === orgInstanceUrl,
+      (org) => org.instanceUrl === orgInstanceUrl
     );
-    if (matchOrgs) {
+    if (matchOrgs.length > 0) {
       this.majorOrgBranch = matchOrgs[0].branch;
       return true;
     }
@@ -244,10 +258,10 @@ export class HardisColors {
     if (vscode.workspace.workspaceFolders) {
       const sfdxHardisConfigFilesPattern = new vscode.RelativePattern(
         vscode.workspace.workspaceFolders[0],
-        `**/.sfdx-hardis*.yml`,
+        `**/.sfdx-hardis*.yml`
       );
       const fileUris = await vscode.workspace.findFiles(
-        sfdxHardisConfigFilesPattern,
+        sfdxHardisConfigFilesPattern
       );
       const orgInstanceUrls = [];
       for (const fileUri of fileUris) {